added fetchall method and updated readme

+++ lab-shelly/controller/cat-controller.js
@@ -7,24 +7,27 @@ const Cat = require('../model/cat');
 module.exports = exports = {};
 
 exports.createItem = function(item) {
+  if(!item) return Promise.reject(createError(400, 'item required'));
   return new Cat(item).save();
 };
 
 exports.fetchItem = function(id){
-  return Cat.findById(id)
-  .catch(err => Promise.reject(createError, 404, err.message));
+  if(!id) return Promise.reject(createError(400, 'id required'));
+  return Cat.findById(id);
+};
+
+exports.fetchAll = function() {
+  return Cat.find();
 };
 
 exports.deleteItem = function(id) {
-  return Cat.findByIdAndRemove(id)
-  .catch(err => Promise.reject(createError(404, err.message)));
+  if(!id) return Promise.reject(createError(400, 'id required'));
+  return Cat.findByIdAndRemove(id);
 };
 
 exports.updateItem = function(id, item) {
-  // if(!schema) return Promise.reject(createError(400, 'schema required'));
   if(!id) return Promise.reject(createError(400, 'id required'));
   if(!item) return Promise.reject(createError(400, 'name and mood required'));
 
-  return Cat.findByIdAndUpdate(id, {name: item.name, mood: item.mood}, {new: true})
-  .catch(err => Promise.reject(createError(404, err.message)));
+  return Cat.findByIdAndUpdate(id, {name: item.name, mood: item.mood}, {new: true});
 };
